requirements.txt
asgiref==3.11.1
certifi==2026.1.4
charset-normalizer==3.4.4
Django==5.0.14
idna==3.11
loguru==0.7.3
markdown-it-py==4.0.0
mdurl==0.1.2
numpy==2.4.2
pandas==3.0.1
pillow==12.1.1
psycopg==3.3.3
psycopg-binary==3.3.3
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
reportlab==4.4.10
requests==2.32.5
rich==14.3.3
schedule==1.2.2
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.3
